First commit
Cripper is used to get information about the files in a directory and to
store that information in a database file. The information in the
database file will then be used to determine which files were added
when, how long they have been there, what files are new since certain
dates and many other uses.