Got the main program working
The run_id is now properly retrieved from the database for use when
inserting records. The loop that goes through the walk() now uses a sub
loops to go through all of the files in each dir and writes them to the
database with their path and the run_id.